feat(autofix): Remove AI from naming and fix crash for old autofix results
