+++ fixtures/js-stubs/autofixStep.tsx
@@ -1,4 +1,4 @@
-import {AutofixStep} from 'sentry/components/events/aiAutofix/types';
+import {AutofixStep} from 'sentry/components/events/autofix/types';
 
 export function AutofixStepFixture(params: Partial<AutofixStep>): AutofixStep {
   return {
