+++ static/app/components/events/autofix/autofixBanner.spec.tsx
